% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boston}
\alias{boston}
\title{Results from the Boston Marathon}
\format{
A data frame with 649378 rows and 7 variables:
\describe{
  \item{display_name}{Runner name}
  \item{age}{Age in years}
  \item{gender}{Male (M) and female (F)}
  \item{seconds}{Time in seconds to complete the marathon}
  \item{overall}{Overall place}
  \item{gender_result}{Place in gender category}
  \item{year}{Year of the race}
  }
}
\usage{
boston
}
\description{
Data retrieved from the Boston-Marathon-Data-Project
https://github.com/adrian3/Boston-Marathon-Data-Project
}
\references{
{Adrian Hanft (2019) https://tread1st.com/boston.html}
}
\keyword{datasets}
