% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dxadata}
\alias{dxadata}
\title{Dual-energy x-ray absorptiometry (DXA) data from a training study}
\format{
A data frame with 80 rows and 101 variables:
\describe{
  \item{participant}{participant identification}
  \item{time}{Time-point of data collection, before (pre) or after (post) 12-weeks of resistance training}
  \item{multiple}{Identification of what leg (L, left; R, right) is associated with multiple-set training (three sets per exercise).}
  \item{single}{Identification of what leg (L, left; R, right) is associated with single-set training (one sets per exercise).}
  \item{sex}{Female/Male}
  \item{include}{Participants included or excluded from data analysis.}
  \item{age}{Participant age}
  \item{height}{Participant height (cm)}
  \item{weight}{Participant weight (kg)}
  \item{BMD.*}{Bone mineral density, grams per centimeter squared (g/cm2) for body except head (body), ribs, pelvis, spine or whole body (whole).}
  \item{fat.*}{Fat mass in grams (g) for left (left_\*) or right (right_\*) side from the whole body (whole), body (except head, body), arm or leg, android and gynoid.}
  \item{lean*.}{Lean mass in grams (g) for left (left_\*) or right (right_\*) side from the whole body (whole), body (except head, body), arm or leg, android and gynoid.}
  \item{BMC.*}{Bone mineral content in grams (g) for left (left_\*) or right (right_\*) side from the whole body (whole), body (except head, body), arm or leg, android and gynoid.}
}
}
\usage{
dxadata
}
\description{
Data from Hammarström et al. (2020)
}
\references{
{Hammarström, D., et al. (2020). "Benefits of higher resistance-training volume are related to ribosome biogenesis." The Journal of Physiology 598(3): 543-565.}
}
\keyword{datasets}
