% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{goldberg}
\alias{goldberg}
\title{Increase in Protein synthesis and muscle wet weight in rat skeletal muscle}
\format{
A data frame with 42 rows and 3 variables:
\describe{
 \item{muscle.ww}{Muscle wet weight, expressed as a percentage of 
 the contra-lateral sham-operated limb}
 \item{leucin.incorp}{Protein synthesis measured as incorporation of labelled
  leucin, expressed as a percentage of the contra-lateral sham-operated limb}
 \item{muscle}{Identification of muscle, plantaris or soleus}
 }
 
 @references {Goldberg, Alfred L. 1968. “Protein Synthesis during Work-Induced
  Growth of Skeletal Muscle.” The Journal of Cell Biology 36(3):653–58}
}
\usage{
goldberg
}
\description{
Data from Goldberg (1968). Muscle growth and incorporation of labelled leucin 
into protein in rat skeletal muscle (plantaris and soleus) after tenotomy of 
the gastrocnemius insertions into the Achilles tendon. Values are presented as
percentage of the contra-lateral sham-operated limb.
}
\keyword{datasets}
