% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hbmass}
\alias{hbmass}
\title{Hemoglobin mass in cyclists}
\format{
A data frame with 46 rows and 4 variables:
\describe{
 \item{id}{Participant identification}
 \item{group}{Grouping, experimental group (heat) and control group (con)}
 \item{time}{Time point, pre and post-intervention}
 \item{hb}{Hbmass (g)}
}
}
\usage{
hbmass
}
\description{
Data from Rønnestad (2021). A group of cyclists trained for five weeks
in a heat chamber (heat group) or in control conditions (con group).
}
\references{
{Rønnestad, B. R. et al. Five weeks of heat training increases haemoglobin mass in elite cyclists. Experimental Physiology 106, 316–327 (2021).}
}
\keyword{datasets}
