% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{heatxc}
\alias{heatxc}
\title{Hemoglobin mass in cross coutry skiers - Effects of heat training}
\format{
A data frame with 50 rows and 26 variables:
\describe{
 \item{participant}{Participant identification}
 \item{group}{Grouping, experimental group (exp) and control group (con)}
 \item{time}{Time point, pre and post-intervention}
 \item{weight}{Body mass (kg)}
 \item{height}{Stature (cm)}
 \item{hb}{Hemoglobin concentration (g/dl)}
 \item{hct}{Hematocrit (vol\%)}
 \item{hbmass}{Hemoglobin mass (g)}
 \item{rbcv}{Red blood cell volume (ml)}
 \item{pv}{Plasma volume (ml)}
 \item{bv}{Blood volume (ml)}
 \item{hf_submax}{Submaximal heart rate (BPM)}
 \item{la_submax}{Submaximal lactate (mml L-1)}
 \item{borg_submax}{Submaximal Borg RPE (Borg 6-20)}
 \item{VO2_submax}{ Submaximal VO2 (ml min-1)}
 \item{lt}{Lactate threshold km h-1}
 \item{lt_VO2}{Lactate threshold as percentage of VO2max (\%)}
 \item{VO2max}{VO2max (ml min-1)}
 \item{VO2max_kg}{Relative VO2max (ml kg-1 min-1)}
 \item{la_max}{Lactate after maximal test (mmol L-1)}
 \item{borg_max}{Borg after maximal test (Borg 6-20)}
 \item{hf_max}{ Heart rate at end of maximal test (BPM)}
 \item{tte}{Time to exhaustion during maximal test (sec)}
 \item{distance_allout}{Distance during all out test (m)}
 \item{borg_allout}{Borg after all out test (Borg 6-20)}
 \item{lactate_allout}{ Lactate after all out test (mmol l-1)}
}
}
\usage{
heatxc
}
\description{
Data from Rønnestad (2022) where a group of cross country skiers completed either
heat suit (n = 13) or control training (n = 12) for five weeks. Hematological variables were
measured before and after the intervention along with performance, VO2max and related variables.
}
\references{
{Rønnestad BR, Lid OM, Hansen J, et al. Heat suit training increases hemoglobin mass in elite cross-country skiers. Scand J Med Sci Sports. 2022; 32: 1089–1098. doi:10.1111/sms.14156}
}
\keyword{datasets}
