% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hypertrophy}
\alias{hypertrophy}
\title{Body composition, strength and molecular data from a resistance training study}
\format{
A data frame with 30 rows and 147 variables:
\describe{
  \item{}{}
  \item{PARTICIPANT}{Participant ID}            
  \item{AGE}{Participants in years}
  \item{BODYMASS_\*}{Body mass (kg) at time-point T\*}T1             
  \item{SQUAT_3RM}{Squat 3 repetition maximum load (kg) at baseline}              
  \item{SQUAT_VOLUME}{Total back-squat training volume during intervention (kg)}            
  \item{GROUP}{Intervention group, participants supplemented with 30 g daily maltodextrin (MALTO), 25 g of whey protein (WP) or graded whey protein (GWP, 25 g/day increasing by 25 g/d per week).}                
  \item{CLUSTER}{Cluster responder grouping based on (a) right leg VL muscle thickness assessed via ultrasound, (b) upper right leg lean soft tissue assessed via DXA, (c) right leg mid-thigh circumference (d) right leg VL mean (type I and type II) muscle fCSA. See Haun et al. (2019)}               
  \item{PERCENT_TYPE_I_T1}{Percentage Type I fibers at baseline (T1)}       
  \item{PERCENT_TYPE_II_T1}{Percentage Type II fibers at baseline (T1)}    
  \item{FAST_CSA_\*}{Type II (FAST) fibers cross sectional area (micrometer^2)}             
  \item{SLOW_CSA_\*}{Type I (SLOW) fibers cross sectional area (micrometer^2)}                 
  \item{FAST_NUCLEI_T1}{Myonuclei per fiber CSA in Type II (FAST)}                          
  \item{SLOW_NUCLEI_T1}{Myonuclei per fiber CSA in Type I (SLOW)}    
  \item{AR_PROTEIN_\*}{Androgen receptor protein content at time-point \*}          
  \item{PROTEASOME_\*}{20S Proteasome Activity at time-point \* (relative fluorometric units)}          
  \item{GLYCOGEN_\*}{Whole muscle tissue glycogen concentration (nmol glycogen/mg wet muscle weight) at time-point \*}            
  \item{CS_\*}{Citrate synthase activity at time-point \*, (mM min-1 mg-1)}                 
  \item{CK_\*}{Creatine kinase activity in blood samples (U L-1) at time-point \*}                  
  \item{TESTOSTERONE_\*}{Testosterone concentration at time-point \* (ng dL-1) }       
  \item{CORTISOL_\*}{Cortisol concentration at time-point \* (ng dL-1) }             
  \item{PAN4EBP1_\*, PANMTOR_\*, PANAMPK_\*, PANP70S6K_\*}{Total protein (PAN) levels of 4EBP-1, mTOR, AMPK and p70S6K1 proteins at time-point \* (AU)}         
  \item{PHOSPHO4EBP1_\*, PHOSPHOMTOR_\*, PHOSPHOAMPK_\*, PHOSPHOP70S6K_\*}{Phosphorylated protein (PHOSPHO) levels of 4EBP-1, mTOR, AMPK and p70S6K1 proteins at time-point \* (AU)}                
  \item{PANPOLYUB_\*}{Ubiquitin-labeled protein levels at time-point \* (AU)}           
  \item{RNA_\*}{Total RNA concentrations at time-point \*, (micro-gram mg-1)}                 
  \item{MGF_T2T1_FOLD_CHANGE}{Mechano growth factor mRNA, fold-change from baseline (T1) to time-point T2}   
  \item{MGF_T3T1_FOLD_CHANGE}{Mechano growth factor mRNA, fold-change from baseline (T1) to time-point T3}    
  \item{MGF_T1}{Baseline mechano growth factor mRNA (AU)}                 
  \item{MSTN_T2T1_FOLD_CHANGE}{Myostatin mRNA, fold-change from baseline (T1) to time-point T2}     
  \item{MSTN_T3T1_FOLD_CHANGE}{Myostatin mRNA, fold-change from baseline (T1) to time-point T3}        
  \item{MSTN_T1}{Baseline myostatin mRNA (AU, expressed per "housekeeping genes")}                  
  \item{RNA45S_T1}{rRNA 45S baseline expression (AU, expressed per "housekeeping genes")}             
  \item{RNA45S_T2T1_FOLD_CHANGE}{rRNA 45S, fold-change from baseline (T1) to time-point T2}
  \item{RNA45S_T3T1_FOLD_CHANGE}{rRNA 45S, fold-change from baseline (T1) to time-point T3}
  \item{DXA_\*_\*}{Dual x-ray absorptiometry lean body mass (LBM) or fat mass (FM)_\*_ from time-point (_\*)}            
  \item{HEIGHT}{Participant stature (cm)}                
  \item{TRAINING_AGE}{Number of years of previous training (years)} 
  \item{VL_\*}{Vastus lateralis thickness at time-point \*, meassured with ultrasound (cm)}
  \item{BICEPS_\*}{Biceps bracchi thickness at time-point \*, meassured with ultrasound (cm)}         
  \item{ECW_\*, ICW_\*, TBW_\*}{Extracellular (ECW), Intracellular (ICW) or total body water (TBW) meassured at time-point \*, (kg)}                
  \item{TOTAL_VOLUME_LOAD_WEEK_\*}{Total training volume at week \* (kg)}
  \item{CALS_WEEK_\*, PROTEIN_WEEK_\*, CHO_WEEK_\*, FAT_WEEK_\*}{Registred energy (CALS, kcal/day), protein (PROTEIN, g/day), carbohydrate (CHO, g/day), fat (FAT, g/day) at week \*.}         
  \item{TMD_\*}{Total mood disturbance (POMS score) at time point \*}                  
  \item{PPT_AVG_\*}{Pressure-to-pain threshold (N) at time point \*}              
}
}
\usage{
hypertrophy
}
\description{
Data from Haun et al. (2019) and Haun et al. (2018)
}
\references{
{Haun, C. T., et al. (2019). "Pre-training Skeletal Muscle Fiber Size and Predominant Fiber Type Best Predict Hypertrophic Responses to 6 Weeks of Resistance Training in Previously Trained Young Men." Front Physiol 10(297).}

{Haun, C. T., et al. (2018). "Effects of Graded Whey Supplementation During Extreme-Volume Resistance Training." Front Nutr 5: 84.}
}
\keyword{datasets}
