% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{injury}
\alias{injury}
\title{Chinese university football injuries}
\format{
A data frame with 800 rows and 19 variables:
\describe{
  \item{Age}{Age in years}
  \item{Height_cm}{Height (cm)}
  \item{Weight_kg}{Weight (kg)}
  \item{Position}{Player position; "Midfielder", "Forward", "Defender", "Goalkeeper"}
  \item{Training_Hours_Per_Week    }{Training volume (Hours per week)}
  \item{Matches_Played_Past_Season }{Number of matches played}
  \item{Knee_Strength_Score        }{Knee strength score}
  \item{Hamstring_Flexibility      }{Hamstring flexibility score}
  \item{Reaction_Time_ms           }{Reaction time (ms)}
  \item{Balance_Test_Score         }{Balance test score}
  \item{Sprint_Speed_10m_s         }{10 m Sprint (seconds)}
  \item{Agility_Score              }{Agility score}
  \item{Sleep_Hours_Per_Night      }{Sleep, Hours per night}
  \item{Stress_Level_Score         }{Stress score}
  \item{Nutrition_Quality_Score    }{Nutrition quality score}
  \item{Warmup_Routine_Adherence}{Adherence to warm-up routine}
  \item{Injury_Next_Season }{Binary classification where injury is defined
   as training/competition-related injury causing ≥7 consecutive days of
    absence, verified by university medical center and coaching staff.}
  \item{BMI}{weight / height^2 }
  }
}
\usage{
injury
}
\description{
The data set contains 800 Chinese university football 
players characteristics and injury occurrence during 
the next season.
}
\details{
Data retreived from Kaggle. Data licensed under CC0: Public Domain.
DOI: https://doi.org/10.34740/kaggle/dsv/12653694
}
\references{
{ChunHong Yuan (2025) University Football Injury 
Prediction Dataset. DOI: 10.34740/KAGGLE/DSV/12653694. Publisher: Kaggle}
}
\keyword{datasets}
