% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{millward}
\alias{millward}
\title{Protein synthesis and Total RNA concentration in rat skeletal muscle}
\format{
A data frame with 52 rows and 3 variables:
\describe{
 \item{group}{group identification, A or B}
 \item{RNA}{Total RNA concentration in rat skeletal muscle, microgram per milligram protein}
 \item{protein_synthesis}{Fractional protein synthesis rate in rat skeletal muscle, Ks day-1}
}
}
\usage{
millward
}
\description{
Data from Millward (1973) to describe the relationship between Total RNA and 
protein synthesis in rat skeletal muscle. Protein synthesis was measured 
using constant intravenous infusion of 14C-labelled tyrosine in rats. Rats 
where either fed a standard diet (10\% protein), starved, or refed after 
starvation with a protein-free diet or control diet. In the data set, two
groups are identified where A represent rats on control diet or refed animals
and B represent starved/fasted animals and animals refed with protein-free
diet.
}
\references{
{Millward, D. J., P. J. Garlick, W. P. T. James, D. O. Nnanyelugo, og J. S. Ryatt. 1973. «Relationship between Protein Synthesis and RNA Content in Skeletal Muscle». Nature 241:204. doi: 10.1038/241204a0.}
}
\keyword{datasets}
